Move serialize and deserialize into keml service
